Refactor into a hints module
