+++ records_mover/db/mysql/load_options.py
@@ -3,7 +3,7 @@ from records_mover.records import RecordsHints
 from records_mover.utils import quiet_remove
 from records_mover.records.hints import cant_handle_hint
 from records_mover.records import DelimitedRecordsFormat
-from records_mover.records.types import (
+from records_mover.records.hints import (
     HintEncoding, HintRecordTerminator,
     HintFieldDelimiter, HintQuoteChar,
     HintQuoting, HintEscape,
