+++ records_mover/db/postgres/copy_options/__init__.py
@@ -1,5 +1,5 @@
 from records_mover.records.load_plan import RecordsLoadPlan
-from records_mover.records.types import RecordsHints
+from records_mover.records.hints import RecordsHints
 from records_mover.records.records_format import DelimitedRecordsFormat
 import logging
 from typing import Set, Tuple, Optional
