+++ records_mover/records/pandas/prep_for_csv.py
@@ -29,7 +29,7 @@ def _convert_series_or_index(series_or_index: T,
                         "string in CSV's format")
             hint_date_format = records_format.hints['dateformat']
             assert isinstance(hint_date_format, str)
-            pandas_date_format = python_date_format_from_hints.get(hint_date_format)
+            pandas_date_format = python_date_format_from_hints.get(hint_date_format)  # type: ignore
             if pandas_date_format is None:
                 cant_handle_hint(processing_instructions.fail_if_cant_handle_hint,
                                  'dateformat',
@@ -48,7 +48,7 @@ def _convert_series_or_index(series_or_index: T,
                         "in CSV's format")
             hint_time_format = records_format.hints['timeonlyformat']
             assert isinstance(hint_time_format, str)
-            pandas_time_format = python_time_format_from_hints.get(hint_time_format)
+            pandas_time_format = python_time_format_from_hints.get(hint_time_format)  # type: ignore
             if pandas_time_format is None:
                 cant_handle_hint(processing_instructions.fail_if_cant_handle_hint,
                                  'timeonlyformat',
